requirements.txt
ansi2html==1.9.1
asgiref==3.7.2
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect2==0.7.0
dash==2.14.2
dash-ag-grid==2.4.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-appconf==1.0.6
django-bootstrap-icons==0.8.7
django-bootstrap4==24.1
django-classy-tags==4.1.0
django-cms==4.1.0rc4
django-comments-xtd==2.9.10
django-contrib-comments==2.2.0
django-crispy-forms==2.1
django-entangled==0.5.4
django-filer==3.1.1
django-filter==23.4
django-formtools==2.4.1
django-fsm==2.8.1
django-import-export==3.3.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-parler==2.3
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-select2==8.1.2
django-tables2==2.7.0
django-treebeard==4.7
django-widget-tweaks==1.5.0
djangocms-admin-style==3.2.6
djangocms-alias==2.0.0rc2
djangocms-attributes-field==3.0.0
djangocms-file==3.0.1
djangocms-frontend==1.2.0
djangocms-text-ckeditor==5.1.5
djangocms-versioning==2.0.0rc1
djangorestframework==3.14.0
docutils==0.20.1
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
Flask==3.0.0
html5lib==1.1
idna==3.6
import-export==0.3.1
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
Markdown==3.5.1
MarkupPy==1.14
MarkupSafe==2.1.3
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysqlclient==2.2.1
nest-asyncio==1.5.8
numpy==1.26.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
plotly==5.18.0
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.0.7
requests==2.31.0
retrying==1.3.4
setuptools==65.5.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
tabulate==0.9.0
tenacity==8.2.3
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.3
unicodedata2==15.1.0
urllib3==2.1.0
webencodings==0.5.1
Werkzeug==3.0.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.17.0
